SCRUM-53 revised gemini implementation
The first iteration as using the old sdk package that was set to be deprecated. Now using up to date sdk. Also implemented was a strict configuration so that gemini only produces a json structure that we can parse under the instructed format. This will make parsing the gemini output and integrating with the exercisedb much more seamless. This is a continuation from the first commit